commit budget management feature

+++ src/main/java/org/codewith3h/finmateapplication/repository/TransactionRepository.java
@@ -1,14 +1,23 @@
 package org.codewith3h.finmateapplication.repository;
 
 import org.codewith3h.finmateapplication.entity.Transaction;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.time.LocalDate;
+import java.util.List;
+import java.util.Optional;
+
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
     Page<Transaction> findByUserId(Integer userId, Pageable pageable);
 
-    Optional<Transaction> findByIdAndUserId(Integer transactionId,  Integer userId);
+    Optional<Transaction> findByIdAndUserId(Integer transactionId, Integer userId);
 
     Page<Transaction> findByUserIdAndTransactionDateBetween(Integer userId, LocalDate startDate, LocalDate endDate, Pageable pageable);
